Harmonize INT_MAX/INT_MIN and KEY_ENTER/KEY_BACKSPACE/FINEANGLES divergent forms
Fixes value/type-form inconsistencies found in the cross-file const
survey. Doesn't consolidate these into a canonical file (they remain
duplicated per-file, same as before) - just makes every copy of each
name textually and semantically identical.

INT_MAX/INT_MIN: m_fixed.rs already used the idiomatic `i32::MAX`/
`i32::MIN` form (from an earlier cleanup pass); 10 other files computed
INT_MAX from a local `__INT_MAX__: i32 = 2147483647 as i32` const
instead (4 of them - m_bbox.rs, p_inter.rs, r_segs.rs, p_mobj.rs - also
computed INT_MIN as `-__INT_MAX__ - 1 as i32`). Switched all of these
to `i32::MAX`/`i32::MIN` directly and removed the now-fully-vestigial
local __INT_MAX__ const from all 11 files (confirmed unused elsewhere).

KEY_ENTER/KEY_BACKSPACE/FINEANGLES: pure literal-spelling drift, same
value everywhere (`13 as i32` vs `13`, `0x7f as i32` vs `127`, `8192 as
i32` vs `8192`). Standardized each to the simplest cast-free form the
outlier file in each group already had - the const's own type
annotation already makes the `as i32` cast redundant.

Verified: 0 build errors, warning diff against base is byte-identical
(these are pure value-representation changes, same runtime values).
Release build + Xvfb boot smoke test confirm the game still runs.

+++ rust-doomgeneric/src/hu_stuff.rs
@@ -41,7 +41,7 @@ pub const false_0: i32 = 0 as i32;
 pub const TICRATE: i32 = 35 as i32;
 pub const MAXPLAYERS: i32 = 4 as i32;
 pub const KEY_ESCAPE: i32 = 27 as i32;
-pub const KEY_ENTER: i32 = 13 as i32;
+pub const KEY_ENTER: i32 = 13;
 pub const KEY_RSHIFT: i32 = 0x80 as i32
     + 0x36 as i32;
 pub const KEY_RALT: i32 = 0x80 as i32
